pyproject.toml
[tool.poetry]
name = "repl_python3_keyboardmovement"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
keyboard = "^0.13.5"
"RPi.GPIO" = "^0.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
